Restructure the project to have the correct org.dataone.xyz structure
