+++ src/main/java/org/dataone/speedbagit/SpeedBagException.java
@@ -1,4 +1,4 @@
-package org.dataone;
+package org.dataone.speedbagit;
 
 
 public class SpeedBagException extends Exception {
