+++ src/main/java/org/dataone/speedbagit/SpeedBagIt.java
@@ -1,4 +1,4 @@
-package org.dataone;
+package org.dataone.speedbagit;
 
 
 import java.nio.charset.StandardCharsets;
