+++ src/test/java/org/dataone/speedbagit/SpeedStreamTest.java
@@ -1,4 +1,4 @@
-package org.dataone;
+package org.dataone.speedbagit;
 
 public class SpeedStreamTest {
 }
